# toolverse-ai-backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TOOLVERSE AI Backend API",
    "version": "1.0.0",
    "description": "Backend-only API for the all-in-one AI tools + file-converter platform. Frontend consumes these endpoints. Auth via Bearer JWT or tv_ API key. File tools use multipart field \"files\"."
  },
  "servers": [
    {
      "url": "/api",
      "description": "API base"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/tools/{slug}": {
      "get": {
        "summary": "Get tool",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/tools/{toolId}/process": {
      "post": {
        "summary": "Process (enqueue job, returns jobId immediately)",
        "parameters": [
          {
            "name": "toolId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "string",
                    "format": "binary"
                  },
                  "text": {
                    "type": "string"
                  },
                  "params": {
                    "type": "string"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued"
          }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "summary": "Job status (poll)",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/jobs/{jobId}/cancel": {
      "post": {
        "summary": "Cancel job",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/jobs/{jobId}/download": {
      "get": {
        "summary": "Download output (owner JWT or ?token= signed token)",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "fileId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File bytes"
          }
        }
      }
    },
...
  },
...
}